pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmanalyze"
version = "0.1.0"
description = "A modular, SOLID Python bot for stock market analysis using LLMs, technical indicators, and news sentiment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "openai>=1.30.0",
    "google-genai>=0.1.0",
    "yfinance>=0.2.40",
    "ta>=0.11.0",
    "pandas>=2.2.2",
    "newsapi-python>=0.2.7",
    "feedparser>=6.0.11",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "redis>=5.0.4",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
